Move announcement logic to the hooks
